Split getServerSideProps into separate file
